requirements.txt
pandas==0.25.1
requests==2.22.0
backports.os==0.1.1
unicodecsv==0.14.1
json5==0.8.5
mysql-connector==2.2.9
beautifulsoup4==4.8.0